+++ src/aqua_diagnostics/core/__init__.py
@@ -2,8 +2,10 @@ from .diagnostic import Diagnostic
 from .util import template_parse_arguments, open_cluster, close_cluster
 from .util import load_diagnostic_config, merge_config_args
 from .util import convert_data_units
+from .output_saver import OutputSaver
 
 __all__ = ['Diagnostic',
            'template_parse_arguments', 'open_cluster', 'close_cluster',
            'load_diagnostic_config', 'merge_config_args',
-           'convert_data_units']
+           'convert_data_units',
+           'OutputSaver']