+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Removed:
 - `aqua.slurm` has been removed.
 
 AQUA core complete list:
+- New OutputSaver class (#1837)
 - Introduce a `Timstat()` module independent from the `Reader()` (#1832)
 - Adapt Catalog Generator to Data-Portfolio v1.3.0 (#1848)
 - Introduction of a internal AQUA data model able to guess coordinates and convert toward required target data convention definition (#1862, #1877)